refactor(types): split molecule schemas into individual modules
