+++ packages/types/src/page/molecules/index.ts
@@ -1,0 +1,13 @@
+export * from "./AnnouncementBar";
+export * from "./ValueProps";
+export * from "./ReviewsCarousel";
+export * from "./ContactForm";
+export * from "./NewsletterSignup";
+export * from "./SearchBar";
+export * from "./MapBlock";
+export * from "./VideoBlock";
+export * from "./FAQBlock";
+export * from "./CountdownTimer";
+export * from "./SocialLinks";
+export * from "./SocialFeed";
+export * from "./SocialProof";
